# public/openapi.yaml
info:
  description: |
    Purpie is an open source social media solution focusing on video oriented communication and content sharing.
    
    The Purpie endpoints are offered to our users to facilitate the use of our technology and allow seamless integrations with their applications.
    For this documentation, you can use the **Api Key** and **Api Secret** which generated from [Purpie](https://purpie.io/) settings page to test the services.

    # Introduction
    Purpie orchestrates various media services and solutions for online meetings, live streamings, on demand video delivery and many more.
    With an agile and hierarchical way of user interaction and content sharing possibilities, Purpie proposes a robust and flexible solution for
    various video communication and distribution needs.

    # Cross-Origin Resource Sharing
    This API features Cross-Origin Resource Sharing (CORS) implemented.
    And that allows cross-domain communication from the browser.
    All responses have a wildcard same-origin which makes them completely public and accessible to everyone, including any code on any site.

    # Authentication
    To make authorized calls on Purpie API operations, you need to generate an API key and secret in the settings page after login to [Purpie](https://purpie.io/).
    
    After generating API credentials, you need to note your `api secret`. You can access this field once after creating. If you lost or forgot your api
...
    
...
    header:
    
...

...
  x-logo:
...
servers:
...
openapi: 3.0.0
paths:
...
  /v1/user-zone/list:
    get:
      operationId: UserZoneController_getCurrentUserZones
      summary: List User Zones
      description: Get the list of current user's zones. It also adds the public zone which is selected but not joined yet, Its userZoneId will be null when the user zone is not found.
      parameters:
        - name: app-subdomain
          required: false
          in: header
          description: Zone subdomain
          schema:
            type: string
      responses:
        '200':
          description: List of current user's zones.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserZoneListResponse'
        '401':
          description: Error thrown when the user has not authorized or JWT is invalid.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: You not authorized to use this route
                  error:
                    type: string
                    example: NOT_SIGNED_IN
      tags:
        - User Zone
  /v1/user-zone/detail:
    get:
      operationId: UserZoneController_getCurrentUserZone
      summary: Get User Zone
      description: Get current user's zone detail, This is deduced from the current subdomain.
      parameters:
        - name: app-subdomain
          in: header
          description: Zone subdomain
          required: true
          schema:
            type: string
        - name: userZoneId
          required: true
          in: path
          description: User Zone Id
          schema: {}
      responses:
        '200':
          description: Get zone detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserZoneListResponse'
        '401':
          description: Error thrown when the user has not authorized or JWT is invalid.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: You not authorized to use this route
                  error:
                    type: string
                    example: NOT_SIGNED_IN
        '404':
          description: Error thrown when the user zone ist not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 404
                  message:
                    type: string
                    example: User Zone not found
                  error:
                    type: string
                    example: ZONE_NOT_FOUND
      tags:
        - User Zone
  /v1/user-zone/detail/{userZoneId}:
    get:
      operationId: UserZoneController_getUserZoneById
      summary: Get User Zone
      description: Get user zone detail by user zone id.
      parameters:
        - name: userZoneId
          required: true
          in: path
          description: User Zone Id
          schema: {}
      responses:
        '200':
          description: Get user zone
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserZoneListResponse'
        '401':
          description: Error thrown when the user has not authorized or JWT is invalid.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: You not authorized to use this route
                  error:
                    type: string
                    example: NOT_SIGNED_IN
        '404':
          description: User Zone not found. Error thrown when user zone is not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 404
                  message:
                    type: string
                    example: User Zone not found
                  error:
                    type: string
                    example: ZONE_NOT_FOUND
      tags:
        - User Zone
  /v1/user-zone/remove/{userZoneId}:
    delete:
      operationId: UserZoneController_deleteUserZoneById
      summary: Delete User Zone
      description: Delete user zone with "userZoneId" parameter.
      parameters:
        - name: userZoneId
          required: true
          in: path
          description: User Zone Id
          schema: {}
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: string
                example: OK
        '401':
          description: Error thrown when the user has not authorized or JWT is invalid.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: You not authorized to use this route
                  error:
                    type: string
                    example: NOT_SIGNED_IN
        '404':
          description: User Zone not found. Error thrown when user zone is not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 404
                  message:
                    type: string
                    example: User Zone not found
                  error:
                    type: string
                    example: ZONE_NOT_FOUND
      tags:
        - User Zone
  /v1/user-zone/remove/{zoneId}:
    delete:
      operationId: UserZoneController_deleteUserZoneByZoneId
      summary: Delete User Zone
      description: Delete user zone with only "zoneId" parameter.
      parameters:
        - name: zoneId
          required: true
          in: path
          description: Zone Id
          schema: {}
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: string
                example: OK
        '401':
          description: Error thrown when the user has not authorized or JWT is invalid.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: You not authorized to use this route
                  error:
                    type: string
                    example: NOT_SIGNED_IN
        '404':
          description: User Zone not found. Error thrown when user zone is not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 404
                  message:
                    type: string
                    example: User Zone not found
                  error:
                    type: string
                    example: ZONE_NOT_FOUND
      tags:
        - User Zone
...
  /v1/zone/remove/{zoneId}:
    delete:
      operationId: ZoneController_deleteZone
      summary: Delete Zone
      description: User deletes a zone belonging to "zoneId". User must have the "canDelete" permission on zone.
      parameters:
        - name: zoneId
          required: true
          in: path
          description: The zone id
          schema: {}
      responses:
        '200':
          description: Zone deleted
          content:
            application/json:
              schema:
                type: string
                example: OK
        '401':
          description: Error thrown when the user has not authorized or JWT is invalid.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: You not authorized to use this route
                  error:
                    type: string
                    example: NOT_SIGNED_IN
        '404':
          description: User Zone not found. Error thrown when user zone is not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 404
                  message:
                    type: string
                    example: User Zone not found
                  error:
                    type: string
                    example: ZONE_NOT_FOUND
      tags:
        - Zone
...
  /v1/video/create:
    post:
      operationId: VideoController_createVideoPost
      summary: Create Video Post
      description: Create a new video post with requested video file and payload.
      parameters: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CreateVideoDto'
      responses:
        '200':
          description: Upload a video file successfully
        '400':
          description: 'Error thrown when the requested video format invalid. Valid formats: 3gpp, mp4, quicktime, webm, x-flv and mpeg'
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 400
                  message:
                    type: string
                    example: Please upload a valid video format
                  error:
                    type: string
                    example: INVALID_VIDEO_FORMAT
        '401':
          description: Error thrown when the user has not authorized or JWT is invalid.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: You not authorized to use this route
                  error:
                    type: string
                    example: NOT_SIGNED_IN
      tags:
        - Video
...
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    ZoneUser:
      type: object
      properties:
        id:
          type: string
        fullName:
          type: string
        email:
          type: string
      required:
        - id
        - fullName
        - email
    ZoneForUserZoneList:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        subdomain:
          type: string
        displayPhoto:
          type: string
        description:
          type: string
        public:
          type: boolean
        createdBy:
          $ref: '#/components/schemas/ZoneUser'
      required:
        - id
        - name
        - subdomain
        - description
        - public
        - createdBy
    UserZoneRoleResponse:
      type: object
      properties:
        roleCode:
          type: string
          enum:
            - OWNER
            - MODERATOR
            - USER
        zoneId:
          type: string
        roleName:
          type: string
        canCreateChannel:
          type: boolean
        canInvite:
          type: boolean
        canDelete:
          type: boolean
        canEdit:
          type: boolean
        canManageRole:
          type: boolean
      required:
        - roleCode
        - zoneId
        - roleName
        - canCreateChannel
        - canInvite
        - canDelete
        - canEdit
        - canManageRole
    UserZoneListResponse:
      type: object
      properties:
        id:
          type: number
        createdOn:
          format: date-time
          type: string
        zone:
          $ref: '#/components/schemas/ZoneForUserZoneList'
        zoneRole:
          $ref: '#/components/schemas/UserZoneRoleResponse'
      required:
        - id
        - createdOn
        - zone
        - zoneRole
...
    CreateVideoDto:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        channelId:
          type: string
        public:
          type: boolean
        videoFile:
          type: string
          format: binary
        allowDislike:
          type: boolean
        allowReaction:
          type: boolean
        allowComment:
          type: boolean
      required:
        - title
        - videoFile
...